TEST: Create tag and newt seed lists
- Create an array of Tag records to be added
- Create an array of Newt records to be added
- Create an array of Tag names to go with each seeded Newt

The sticky point is associating Tags with Newts.

First I tried creating model/instance methods to

- look up a single Tag IDs by Tag name
- look up an array of Tag IDs by name
- use the association-made `Newt.addTagge` and `Newt.addTagges` setters to request Tag ID lookups and create the join table rows.

I got the tag lookups working, but then I hit the following error before the Tags could be added to the Newts:

> Unhandled rejection Error: ConnectionManager.getConnection was called after the connection manager was closed!

Then it occurred to me that there must be some better way to do this in Sequelize, and research led me to some examples using`include`. Of course, none of these examples use a many-to-many association. I tried it anyway, but couldn't get it to work. The `include` clause just gets completely ignored — no error, just nothing. It doesn't show up in the query log at all.

So, although the `bulkCreate` query works fine on Tags and Newts, there doesn't seem to be a way to associate records via `include` while using this method. Recommended workarounds seem to be

a. make a loop of individual `create` queries, each with its own `include`, or
b. something something mumble transactions.